Refactor rule providers to centralize dynamic filtering logic
